Adds `evaluate_at_x` method to `PartialSecret`
This way we can reuse the computational work we've done if for some reason we
want to evaluate the same set of interpolated points at value other than
`Gf256::zero()`.

As noted, a slight sacrifice to efficiency was made when implementing this
function, in order to reduce the `PartialSecret` size, and increase
precomputation in the standard case of evaluating at `Gf256::zero()`.