Add sample sss module partial secret recovery fns
These should be considered exemplary at this point, but I wanted to start to
flesh out a higher-level way to interact with the `PartialSecret` struct.
Besides more conceptual changes in terms of how to make this interface more
user-friendly, I think the validation needs to be DRYed out, and the error
handling refined.

In particular, all the functions that follow `begin_partial_secret_recovery` are
basically analogues to methods, and it feels like it would be nicer to call them
as such instead of as functions.

Mostly, I'm unsure of how the repository maintainers would like such an
interface to work, so only took my best jab at fleshing this out.