Store redundant PartialSecret data in Recovery
`threshold` and `ids` are no longer stored in `PartialSecret`, reducing the size
of a `Recovery` object by up to 33% (as the number of shares interpolated
grows).

Optimizing for speed (since we will call methods on `PartialSecrets` much more
often than functions in `validate`), the `ids` field of `Recovery` is now of
type `Vec<Gf256>`.

`PartialSecret` no longer holds all its own state and relies on that state being
managed from the outside. While this is not ideal for using `PartialSecret`
independently of `Recovery`, I feel comfortable tying the two classes. I have
added additional assertions to `PartialSecret` to catch errors.