PartialSecrets to BarycentricWeights
Having crippled `PartialSecrets` sufficiently from its original form, I realized
it was best to carry this process out as far as possible, and make this even
more explicit. Storing an `secret: Option<u8>` didn't make sense when it no
longer stored `ids` and `thresholds` and was dependent on outside forces
providing that information, so it could know when to compute the secret. The
`compute_secret` method, now `evaluate_at_zero`, it's own function in the
`lagrange` module, also seemed out of place.

Now `threshold` doesn't needed to be passed to `BarycentricWeights`, and the
type is more accurately described by its name. An outside force is still
responsible for managing the `ids`, and when `threshold` shares have been
interpolated, it will need to call `evaluate_at_zero` with the
`BarycentricWeights` and `ids` (`Recovery` now does this).

The last change to `Recovery` was to make it hold its own `secret` (it
automatically computes this secret as soon as threshold shares have been
interpolated via calls to `new` and `update`). Thus `get_secret` will be faster
because we're just unwrapping a `Some(Vec<u8>)` and rewrapping it in `Ok`,
instead of needing to construct that `Vec<u8>` by iterating over `Option<u8>`s
we need to unwrap and `collect`.